requirements.txt
# primary
pytesseract==0.3.10
pandas==1.5.1
spacy==3.4.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.4.1/en_core_web_md-3.4.1-py3-none-any.whl

# sub-dependancies
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
cymem==2.0.7
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.9
numpy==1.23.4
packaging==21.3
pathy==0.6.2
Pillow==9.3.0
preshed==3.0.8
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
smart-open==5.2.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
thinc==8.1.5
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
